PR #9 still left one edge case unhandled: when staging.cars holds
two rows with text-different but BIGINT-equal external_ids (e.g.
'7012961' and '07012961' from leading-zero variants in a URL slug),
DISTINCT ON (s.external_id) treats them as different and keeps both.
Both rows get cast to BIGINT 7012961 in the SELECT, so the fact table
ends up with two rows sharing the same primary key, and
external_id.unique fails with 'column external_id has 1 non-unique
values' (matching the exact error shape the user reported).

Fix: dedupe on s.external_id::BIGINT (and order by the same), so
text-different but numerically-equal ids collapse into a single row
before they reach the fact's PK column.

Verified:
- Reproduction (raw.cars rows with external_id='7012961' and
  '07012961'): without this fix, make full-refresh fails with
  '1 non-unique values' -- the exact user error. With this fix,
  full-refresh finishes 11/58 and the 7012961 row is unique.
- Backfill window scenario (PR #9 regression test) still passes:
  Jan-only backfill of an external_id with Jan and Mar rows keeps
  the in-window Jan row (price=500000, color=Red).
- make smoke-import-historical still ALL PASS.

The text regex filter ('^\d+$') stays on the text value -- it's a
cheap rejection of garbage before we attempt the BIGINT cast.

+++ bruin_pipeline/assets/facts/fact_car_listings.sql
@@ -155,21 +155,28 @@ columns:
 
 -- staging.cars is keyed on link (its merge primary_key), so the same
 -- listing can appear under multiple URL variants (trailing slash, query
--- string, ...). external_id is the fact-table PK -- collapse any
--- in-window duplicates here, keeping the freshest row per id (latest
--- updated_at, then scraped_at, then link as a deterministic
--- tiebreaker). The updated_at window filter MUST live inside the CTE:
--- if it lived in the outer WHERE, a backfill of an old window could
--- pick an out-of-window newer row that the outer filter then drops,
--- silently losing the listing from the rebuild.
+-- string, leading zeros on the id, ...). external_id is the fact-table
+-- PK as BIGINT -- collapse any in-window duplicates here, keeping the
+-- freshest row per id (latest updated_at, then scraped_at, then link
+-- as a deterministic tiebreaker).
+--
+-- Dedup MUST happen on the BIGINT-cast value, not the raw text: e.g.
+-- '7012961' and '07012961' are different strings but the same id, and
+-- text-only DISTINCT ON keeps both, leading to a duplicate-PK failure
+-- in the fact's external_id.unique check.
+--
+-- The updated_at window filter also lives inside the CTE: if it lived
+-- only in the outer WHERE, a backfill of an old window could pick an
+-- out-of-window newer row that the outer filter then drops, silently
+-- losing the listing from the rebuild.
 WITH staging_dedup AS (
-    SELECT DISTINCT ON (s.external_id) s.*
+    SELECT DISTINCT ON (s.external_id::BIGINT) s.*
     FROM staging.cars s
     WHERE s.external_id IS NOT NULL
       AND s.external_id ~ '^\d+$'
       AND s.updated_at >= '{{ start_timestamp }}'
       AND s.updated_at <  '{{ end_timestamp }}'
-    ORDER BY s.external_id, s.updated_at DESC, s.scraped_at DESC, s.link
+    ORDER BY s.external_id::BIGINT, s.updated_at DESC, s.scraped_at DESC, s.link
 )
 SELECT
     s.external_id::BIGINT                       AS external_id,
